requirements.txt
# 核心依赖
selenium>=4.0.0
pytest>=7.0.0
pyyaml>=6.0

# 可选依赖（用于报告生成）
# allure-pytest>=2.0.0
# pytest-html>=3.0.0
